cpf-soc-integration/src/ml-models/cpf_risk_model.py
#!/usr/bin/env python3
"""
CPF Risk Model
Behavioral pattern detection for security vulnerabilities
"""

import tensorflow as tf
import numpy as np

class CPFRiskModel(tf.keras.Model):
    """CPF Risk Scoring Model"""
    
    def __init__(self):
        super().__init__()
        self.lstm = tf.keras.layers.LSTM(128, return_sequences=True)
        self.attention = tf.keras.layers.MultiHeadAttention(
            num_heads=8, key_dim=64)
        self.dense1 = tf.keras.layers.Dense(256, activation='relu')
        self.dropout = tf.keras.layers.Dropout(0.3)
        self.output_layer = tf.keras.layers.Dense(100, 
            activation='sigmoid')  # 100 CPF indicators
    
    def call(self, inputs, training=False):
        x = self.lstm(inputs)
        x = self.attention(x, x)
        x = self.dense1(x)
        if training:
            x = self.dropout(x)
        return self.output_layer(x)

if __name__ == "__main__":
    print("CPF Risk Model v1.0")
    model = CPFRiskModel()
    print("Model initialized successfully")
