cpf-soc-integration/src/requirements.txt
pandas
numpy
requests
scikit-learn
redis
pyyaml
